Moved components out of legacy components folder
