+++ src/app/Auth/Auth.js
@@ -1,13 +1,13 @@
 import React from 'react';
 
-import LayoutBootstrap from '../../LayoutBootstrap/LayoutBootstrap';
+import LayoutBootstrap from '../LayoutBootstrap/LayoutBootstrap';
 
 // for unit test, need this default value
 let imageUrl = '/static/google_signin.png';
 
 // TODO: Add loader to webpack config files
 if (process.env.BROWSER) {
-  imageUrl = require('file!../../static/google_signin.png');
+  imageUrl = require('file!./static/google_signin.png');
 }
 
 export default class Auth extends React.Component {
