+++ src/app/Auth/actions/actionCreators.js
@@ -1,23 +1,7 @@
 import fetch from 'isomorphic-fetch';
 import {
-  UPDATE_LOCALE,
   FETCH_USER,
 } from '../constants/actionTypes';
-
-export function setState(state) {
-  return {
-    type: 'SET_STATE',
-    state,
-  };
-}
-
-export function filterEntries(input) {
-  return {
-    meta: { remote: true },
-    type: 'FILTER_ENTRIES',
-    input,
-  };
-}
 
 function fetchUserSuccess(user) {
   return {
@@ -47,12 +31,5 @@ export function fetchAccountDetails() {
     )
     .then(user => dispatch(fetchUserSuccess(user)))
     .catch(e => dispatch(fetchUserFailure(e)));
-  };
-}
-
-export function updateLocale(payload) {
-  return {
-    type: UPDATE_LOCALE,
-    payload,
   };
 }
