+++ src/app/Auth/constants/actionTypes.js
@@ -1,0 +1,1 @@
+export const FETCH_USER = 'FETCH_USER';
