+++ src/app/Auth/reducers/reducers.js
@@ -1,6 +1,6 @@
 import {
   FETCH_USER,
-} from '../../constants/actionTypes';
+} from '../constants/actionTypes';
 
 export default function (state = {}, action) {
   switch (action.type) {
