+++ src/app/CustomNavBar/CustomNavBarContainer.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { connect } from 'react-redux';
 
 import CustomNavBar from './components/CustomNavBar';
-import { updateLocale } from '../actions/action_creators';
+import { updateLocale } from './actions/action_creators';
 
 function CustomNavBarContainer(props) {
   return (
