+++ src/app/CustomNavBar/actions/action_creators.js
@@ -1,0 +1,9 @@
+import { UPDATE_LOCALE } from '../constants/actionTypes';
+
+export function updateLocale(payload) {
+  return {
+    type: UPDATE_LOCALE,
+    payload,
+  };
+}
+
