+++ src/app/CustomNavBar/constants/actionTypes.js
@@ -1,3 +1,1 @@
 export const UPDATE_LOCALE = 'UPDATE_LOCALE';
-
-export const FETCH_USER = 'FETCH_USER';
