--- src/app/containers/NomatchContainer.js
@@ -1,13 +1,0 @@
-import React from 'react';
-
-function NomatchContainer() {
-  return (
-    <div>
-      <p>Page Not found</p>
-    </div>
-  );
-}
-
-NomatchContainer.title = 'backToHome';
-
-export default NomatchContainer;
