--- src/app/reducers/Filter/filter.js
@@ -1,14 +1,0 @@
-import { Map } from 'immutable';
-
-function setState(state, newState) {
-  return state.merge(newState);
-}
-
-export default function (state = new Map(), action) {
-  switch (action.type) {
-    case 'SET_STATE':
-      return setState(state, action.state);
-    default:
-      return state;
-  }
-}
